# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Luna Athletics API",
        "description": "API para gestión de productos y categorías de Luna Athletics",
        "contact": {
            "email": "admin@lunaathletics.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/categorias": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener todas las categorías",
                "operationId": "9bd4c5d463ba7b15843eebef947ed231",
                "responses": {
                    "200": {
                        "description": "Lista de categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Crear una nueva categoría",
                "operationId": "33561abcf774ce0a7d668f231512afd9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCategoriaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Categoria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener una categoría específica",
                "operationId": "0ddbcef186a80938088c691755aee7fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Categoria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Actualizar una categoría",
                "operationId": "3cb64fbc825d55fb192dc610c8ce7a48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoriaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Categoria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Eliminar una categoría",
                "operationId": "5c8248a20636e938f3e6b20513b1d664",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoría eliminada exitosamente"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "53a582135e8459967c18d72e451ad97f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Las credenciales son incorrectas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "8e486c49807eea53b247ac95f493a034",
                "responses": {
                    "200": {
                        "description": "Logout exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Categoria": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "type": "string",
                        "example": "categoria"
                    },
                    "atributos": {
                        "properties": {
                            "nombre": {
                                "type": "string",
                                "example": "Ropa Deportiva"
                            },
                            "descripcion": {
                                "type": "string",
                                "example": "Descripción de la categoría"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CategoriaCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Categoria"
                        }
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "attributes": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    }
                                },
                                "type": "object"
                            },
                            "token": {
                                "type": "string",
                                "example": "1|randomtoken123"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "LogoutResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Cierre de sesión exitoso."
                    }
                },
                "type": "object"
            },
...
            "LoginRequest": {
                "required": [
                    "correo",
                    "contraseña",
                    "dispositivo"
                ],
                "properties": {
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "Edward@example.com"
                    },
                    "contraseña": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "dispositivo": {
                        "type": "string",
                        "example": "windows"
                    }
                },
                "type": "object"
            },
            "StoreCategoriaRequest": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Ropa Deportiva"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Categoría para ropa deportiva",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateCategoriaRequest": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Ropa Deportiva Actualizada"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción actualizada",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}